refactor: use correct naming convention for frame utilities
