+++ packages/dom/src/utilities/index.ts
@@ -50,5 +50,5 @@ export type {Transform} from './transform/index.ts';
 
 export {generateUniqueId} from './misc/generateUniqueId.ts';
 
-export {getFrameElement} from './frame/get-frame-element.ts';
-export {getFrameOffset} from './frame/get-frame-offset.ts';
+export {getFrameElement} from './frame/getFrameElement.ts';
+export {getFrameOffset} from './frame/getFrameOffset.ts';
